requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-xml==2.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
zipp==3.18.1
